--- common/web/input-processor/.gitignore
@@ -1,18 +1,0 @@
-# Build products
-src/environment.inc.ts
-
-# Legacy build folders.
-output/
-build/
-embedded/
-
-# Current build output folder (matches standard node publish location).
-dist/
-
-# Other local files.
-node_modules/
-unit_tests/modernizr.js
-source/environment.inc.ts
-.idea/**/*.xml
-*.iml
-*.kpj.user
