--- common/web/input-processor/README.md
@@ -1,15 +1,0 @@
-# Keyman:  Keyboard Processor Module
-The Original Code is (C) 2017-2020 SIL International
-
-The Keyboard Processor module is an internal component of KeymanWeb, seen within this repo at /web/.
-
-## Minimum Compilation Requirements
-
-* A local installation of [Node.js](https://nodejs.org/) v8.9+.
-	* Builds will call `npm install` to automatically install further necessary build dependencies.
-
-	* Linux users can update to LTS version of nodejs by following instructions on the [NodeSource Distributions](https://github.com/nodesource/distributions#table-of-contents) page.
-
-**********************************************************************
-
-The build script may be found at src/build.sh.
