--- common/web/input-processor/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "@keymanapp/input-processor",
-  "description": "The core text and prediction processing engine for KeymanWeb",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/keymanapp/keyman.git"
-  },
-  "keywords": [
-    "input",
-    "languages",
-    "keyboards"
-  ],
-  "author": "SIL International",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/keymanapp/keyman/issues"
-  },
-  "homepage": "https://github.com/keymanapp/keyman#readme",
-  "devDependencies": {
-    "@keymanapp/kmc": "*",
-    "@keymanapp/resources-gosh": "*",
-    "mocha": "^10.0.0",
-    "mocha-teamcity-reporter": "^4.0.0",
-    "typescript": "^5.4.5"
-  },
-  "scripts": {
-    "test": "gosh ./test.sh"
-  },
-  "dependencies": {
-    "@keymanapp/keyboard-processor": "*",
-    "@keymanapp/lexical-model-layer": "*",
-    "@keymanapp/models-types": "*",
-    "@keymanapp/keyman-version": "*",
-    "@keymanapp/web-utils": "*",
-    "eventemitter3": "^5.0.0"
-  },
-  "type": "module",
-  "main": "./build/obj/index.js",
-  "types": "./build/obj/index.d.ts",
-  "exports": {
-    ".": {
-      "es6-bundling": "./src/index.ts",
-      "default": "./build/obj/index.js"
-    },
-    "./lib": {
-      "types": "./build/lib/index.d.ts",
-      "import": "./build/lib/index.mjs"
-    },
-    "./obj/*.js": "./build/obj/*.js"
-  }
-}
