--- common/web/input-processor/src/index.ts
@@ -1,10 +1,0 @@
-/// <reference path="../../../predictive-text/src/worker-interface.d.ts" />
-
-export * from './corrections.js';
-export * from './correctionLayout.js';
-export { default as InputProcessor } from './text/inputProcessor.js';
-export { default as ContextWindow } from './text/contextWindow.js';
-export { default as ModelSpec } from './text/prediction/modelSpec.js';
-export { default as LanguageProcessor, StateChangeEnum } from './text/prediction/languageProcessor.js';
-export { default as PredictionContext } from './text/prediction/predictionContext.js';
-export { TranscriptionCache } from './transcriptionCache.js';
