--- common/web/input-processor/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "extends": "../tsconfig.kmw-main-base.json",
-  "compilerOptions": {
-    "baseUrl": "./",
-    // Necessary for ES5 iteration of Map.keys().
-    "downlevelIteration": true,
-    "outDir": "build/obj/",
-    "tsBuildInfoFile": "build/obj/tsconfig.tsbuildinfo",
-    "rootDir": "./src"
-  },
-  "references": [
-    { "path": "../utils" },
-    { "path": "../keyboard-processor" },
-    { "path": "../../predictive-text/tsconfig.all.json" },
-  ],
-  // The extra .d.ts is necessary to avoid issues with lack of a Worker type definition when not including
-  // the standard DOM in the TS "lib" configuration.
-  "include": ["./src/**/*.ts", "../../predictive-text/src/worker-interface.d.ts"]
-}
