+++ tsconfig.base.json
@@ -26,7 +26,6 @@
 
     "paths": {
       "@keymanapp/common-types": ["./common/web/types/src/main"],
-      "@keymanapp/input-processor": ["./common/web/input-processor/src"],
       "@keymanapp/keyboard-processor": ["./common/web/keyboard-processor/src"],
       "@keymanapp/keyman": ["./web" ],
       "@keymanapp/models-types": ["./common/models/types"],
