+++ web/README.md
@@ -82,9 +82,7 @@ title: Dependency Graph
 graph TD;
     OSK["/web/src/engine/osk"];
     KP["@keymanapp/keyboard-processor<br>(/common/web/keyboard-processor)"];
-    IP["@keymanapp/input-processor<br>(/common/web/input-processor)"];
     OSK-->KP;
-    IP-->KP;
     WebUtils["@keymanapp/web-utils<br>(/common/web/utils)"];
     KP---->WebUtils;
     Wordbreakers["@keymanapp/models-wordbreakers<br>(/common/models/wordbreakers)"];
@@ -95,7 +93,6 @@ graph TD;
     LMWorker-->Wordbreakers;
     LMLayer["@keymanapp/lexical-model-layer<br>(/common/predictive-text)"];
     LMLayer-->LMWorker;
-    IP-->LMLayer;
     Gestures["@keymanapp/gesture-recognizer<br>(/common/web/gesture-recognizer)"];
     Gestures-->WebUtils;
 
@@ -110,7 +107,6 @@ graph TD;
     Fully headless components`"]
         direction LR
         KP;
-        IP;
         WebUtils;
         PredText;
         Gestures;
@@ -123,22 +119,20 @@ graph TD;
         Elements["/web/src/engine/element-wrappers"];
         Elements-->KP;
         KeyboardCache["/web/src/engine/package-cache"];
-        KeyboardCache-->IP;
-        KeyboardCache-->Configuration;
+        KeyboardCache-->Interfaces;
         DomUtils["/web/src/engine/dom-utils"];
         DomUtils-->WebUtils;
         DomUtils-->KP;
         OSK-->DomUtils;
-        OSK---->IP;
         OSK-->Gestures;
-        Configuration["/web/src/engine/interfaces"];
-        Configuration-->OSK;
+        Interfaces["/web/src/engine/interfaces"];
+        Interfaces-->KP;
+        OSK-->Interfaces;
         CommonEngine["/web/src/engine/main"];
-        CommonEngine-->Configuration;
+        CommonEngine-->Interfaces;
         CommonEngine-->Device;
         CommonEngine-->KeyboardCache;
         CommonEngine-->OSK;
-        CommonEngine-->IP;
         Attachment["/web/src/engine/attachment"];
         Attachment-->DomUtils;
         Attachment-->Elements;
