+++ web/ci.sh
@@ -86,7 +86,6 @@ if builder_start_action test; then
   # No --reporter option exists yet for the headless modules.
 
   "$KEYMAN_ROOT/common/web/keyboard-processor/build.sh" test $OPTIONS
-  "$KEYMAN_ROOT/common/web/input-processor/build.sh" test $OPTIONS
   "$KEYMAN_ROOT/common/web/gesture-recognizer/test.sh" $OPTIONS
 
   ./build.sh test $OPTIONS
