+++ web/src/engine/interfaces/src/index.ts
@@ -1,2 +1,6 @@
 export { default as PathConfiguration } from './pathConfiguration.js';
 export { PathOptionDefaults, PathOptionSpec } from './optionSpec.interface.js';
+export { ModelSpec } from './modelSpec.js';
+export { default as PredictionContext } from './prediction/predictionContext.js';
+export { LanguageProcessorEventMap, StateChangeEnum, ReadySuggestions } from './prediction/languageProcessor.interface.js';
+export { OSKResourcePathConfiguration } from './oskResourcePathConfiguration.interface.js';
