+++ web/src/engine/interfaces/src/oskResourcePathConfiguration.interface.ts
@@ -1,5 +1,5 @@
 // Used by the OSK module.
-export default interface OSKResourcePathConfiguration {
+export interface OSKResourcePathConfiguration {
   readonly fonts: string,
   readonly resources: string
 }
