+++ web/src/engine/main/build.sh
@@ -13,11 +13,14 @@ SUBPROJECT_NAME=engine/main
 # ################################ Main script ################################
 
 builder_describe "Builds the Keyman Engine for Web's common top-level base classes." \
-  "@/common/web/input-processor build" \
+  "@/common/web/keyman-version" \
+  "@/common/web/keyboard-processor" \
+  "@/common/predictive-text" \
   "@/web/src/engine/interfaces build" \
   "@/web/src/engine/device-detect build" \
   "@/web/src/engine/package-cache build" \
   "@/web/src/engine/osk build" \
+  "@/developer/src/kmc-model test" \
   "clean" \
   "configure" \
   "build" \
