+++ web/src/engine/main/src/contextManagerBase.ts
@@ -1,7 +1,7 @@
 import { EventEmitter } from 'eventemitter3';
 import { ManagedPromise, type Keyboard, type KeyboardInterface, type OutputTarget } from '@keymanapp/keyboard-processor';
 import { StubAndKeyboardCache, type KeyboardStub } from 'keyman/engine/package-cache';
-import { PredictionContext } from '@keymanapp/input-processor';
+import { PredictionContext } from 'keyman/engine/interfaces';
 import { EngineConfiguration } from './engineConfiguration.js';
 
 interface EventMap {
