+++ web/src/engine/main/tsconfig.json
@@ -11,7 +11,6 @@
   "include": [ "src/**/*.ts" ],
 
   "references": [
-    { "path": "../../../../common/web/input-processor" },
     { "path": "../device-detect" },
     { "path": "../osk" },
     { "path": "../package-cache" },
