+++ web/src/engine/osk/build.sh
@@ -13,8 +13,9 @@ SUBPROJECT_NAME=engine/osk
 # ################################ Main script ################################
 
 builder_describe "Builds the Keyman Engine for Web's On-Screen Keyboard package (OSK)." \
-  "@/common/web/input-processor build" \
+  "@/common/web/keyboard-processor build" \
   "@/common/web/gesture-recognizer build" \
+  "@/web/src/engine/interfaces build" \
   "@/web/src/engine/dom-utils build" \
   "@/web/src/engine/events build" \
   "clean" \
