+++ web/src/engine/osk/src/index.ts
@@ -8,7 +8,6 @@ export { BannerController } from './banner/bannerController.js';
 // Is referenced by at least one desktop UI module.
 export { FloatingOSKCookie as FloatingOSKViewCookie } from './views/floatingOskCookie.js';
 export { default as VisualKeyboard } from './visualKeyboard.js';
-export { type default as OSKResourcePathConfiguration } from './config/oskResourcePathConfiguration.interface.js';
 export { type default as ViewConfiguration } from './config/viewConfiguration.js';
 export { type KeyElement } from './keyElement.js';
 export { type default as OSKBaseKey } from './keyboard-layout/oskBaseKey.js';
@@ -25,6 +24,9 @@ export { ParsedLengthStyle } from './lengthStyle.js';
 
 export { gestureSetForLayout, DEFAULT_GESTURE_PARAMS } from './input/gestures/specsForLayout.js'
 
-// PredictionContext is exported from input-processor, not the OSK.
+export { PredictionContext } from 'keyman/engine/interfaces';
+
+export * from './corrections.js';
+export * from './correctionLayout.js';
 
 // More things will likely need to be added.
