+++ web/src/engine/osk/src/input/gestures/browser/multitap.ts
@@ -4,7 +4,7 @@ import VisualKeyboard from '../../../visualKeyboard.js';
 import { ActiveSubKey, ActiveKey, KeyDistribution, ActiveKeyBase } from '@keymanapp/keyboard-processor';
 import { GestureSequence, GestureStageReport } from '@keymanapp/gesture-recognizer';
 import { GestureHandler } from '../gestureHandler.js';
-import { distributionFromDistanceMaps } from '@keymanapp/input-processor';
+import { distributionFromDistanceMaps } from 'keyman/engine/osk';
 import Modipress from './modipress.js';
 import { keySupportsModipress } from '../specsForLayout.js';
 import { GesturePreviewHost } from '../../../keyboard-layout/gesturePreviewHost.js';
