+++ web/src/engine/package-cache/src/modelCache.ts
@@ -1,4 +1,4 @@
-import { ModelSpec } from "@keymanapp/input-processor";
+import { ModelSpec } from 'keyman/engine/interfaces';
 
 export default class ModelManager {
   // Tracks registered models by ID.
