+++ web/src/engine/package-cache/tsconfig.json
@@ -11,7 +11,6 @@
   "include": [ "src/**/*.ts" ],
 
   "references": [
-    { "path": "../../../../common/web/input-processor" },
     { "path": "../interfaces" }
   ]
 }
