+++ web/src/tools/testing/recorder/tsconfig.json
@@ -13,7 +13,6 @@
     { "path": "../../../../../common/web/keyman-version" },
     { "path": "../../../../../common/web/utils" },
     { "path": "../../../../../common/web/keyboard-processor" },
-    { "path": "../../../../../common/web/input-processor" },
     { "path": "../../../../../common/web/recorder" },
     { "path": "../../../../../common/web/lm-message-types" },
     { "path": "../../../../../web/src/app/browser" }
